Improvements to fixLieutenants()
Change refs from an iterator to a list, so we actually go through
the pages we should (it gets used twice in fixLieutenants()).
Remove the hard-coded upper limit on items to look for in a Lt
page.
Add TODO noting that not all items that reference faction Lts
have a power that affects all those Lts.