Xref: Sort LTs in Challenge Jobs Table
On the individual Area pages, it's nice to have the Lts in
the order they appear in the image. If you're looking at the
Challenge Jobs Table, though, there are no images, and we want
it to be easy to compare the various jobs. Sorting Lts by rarity
makes this easier.
So this change sorts the Lts on each row so all the Common ones
appear first, then the Uncommon, and finally the Rares.
It might be nice to then sort them alphabetically, but that is
not done here.