Add fix_drop_for_params.py
This script alters any "for" parameter passed to the Drop
template. It adds the link notation '[[...]]' to it. The
intent is that we can the change the template to stop it
doing the same.
There are a couple of special-cases:
- if the page being linked to is not a valid page title
  (this usually occurs if it already has the link notation).
- if the page being linked to doesn't exist, unless it is
  one of the four "Medallion" pages (we currently have a few
  places where "for" is used like this. The Medallion pages
  are expected to exist at some point, though).

I expect that this script will be run just once, around the
time that the Drop template is changed, but it may be useful
to generalise it in the future.