Insert_new_area: Also update Achievements page
Only changes the achievements that are already present.
They use area numbers rather than names, and the numbers
are changed by the insertion.
Addition of new achievements for 5-starring the area and
for killing the new boss is left as a manual task (partly
because there is a lot of information needed to create
those new achievements).