Xref: Add new _remove_implicit_categories() method
Move the code that was in _fix_item() and merge in the new
code from fix_insignia(). In reality, this should probably
be called from many more places. Moving forward, it may also
be possible to read the template pages to determine which
categories are added, rather than relying on the code here
matching the template content.

Note that the optional template parameter currently does
nothing. Not sure at the moment whether I want to use it
or not.