Xref: Some Lts are made in the Tech Lab
Once upon a time, LTs made in the Tech Lab were not available
anywhere else. Some have since effectivelyd disappeared from the
Tech Lab and reappeared in crates, so we need to check sources even
for Tech Lab LTs.

I noticed that Berserker didn't list all the crates he came in,
and this turned out to be the roor cause.