Tables: Improve History parsing for secret jobs dates
Mostly loosening up DATE_RE.
At some point in 2017, months started being spelled out in full, but
the RE was looking for 3-letter month abbreviations (in fact, there were
some uses of "Sept" before that it was also missing).
We occasionally have a month and year but not day-of-the-month.

For the most recent secret jobs (London), the two areas were opened on
different dates, so change the date to a list.