Xref: Make cat Event Lts more meaningful
As it stands, this category is applied to any Lt that is a
reward from an Event. Players, though, when they speak of
"Event Lts", mean that main event rewards, Lts that are *only*
available as the main prize of an event or from the Super Crate.
This change makes the category reflect that meaning, except that
it doesn't remove the category from existing pages.